misc nx-X11/programs/Xserver/{dix,hw/nxagent,include,mi,xfixes}/: backport 73e2383b7350723256894c5076d0c731aec5cd1f even more fuzzily.
This tackles the removal of clientClipType in GC code.

Because the nxagent DDX deviated so much from Xnest, it isn't clear to me
whether we really can switch to using the only supported clip type of Regions
that easily. For instance, nxagent extends its private GC struct with a pixmap
value that is used throughout the code, while Xnest never did this. We might be
in for some trouble, but for now let's hope that Regions are all we need (and
Bitmaps to Region conversions.)

After this commit, there are still places referencing clientClipType in:
  - fb
  - nxagent
  - mi
  - render
code.

We don't care about render for now, but the other three places will need to be
thorougly modified to really get rid of it.