nxcomp/src/Log.cpp: add PID to thread ID output if requested.
The default function we use (if a thread name is not specified
explicitly) is pthread_self(). This function returns a number that is
guaranteed to be unique for each thread within a process, but this
assertion doesn't hold globally. Hence only using the thread ID is
ambiguous when logging from multiple processes.